* Fire event when remapping completes * Organize API - move the marker interfaces for power items that should not be part of public API back into the mod proper (they are used internally in a way that should not be publically exposed, ping me on irc if you want an explanation).
