+++ src/main/java/moze_intel/projecte/api/event/EMCRemapEvent.java
@@ -1,0 +1,10 @@
+package moze_intel.projecte.api.event;
+
+import cpw.mods.fml.common.eventhandler.Event;
+
+/**
+ * Fired after all EMC values are recalculated
+ * This event is not cancelable, and has no result.
+ * This event is fired on MinecraftForge#EVENT_BUS
+ */
+public class EMCRemapEvent extends Event {}
