+++ src/main/java/moze_intel/projecte/api/proxy/ITransmutationProxy.java
@@ -1,4 +1,4 @@
-package moze_intel.projecte.api;
+package moze_intel.projecte.api.proxy;
 
 import java.util.UUID;
