+++ src/main/java/moze_intel/projecte/gameObjs/blocks/Pedestal.java
@@ -3,7 +3,7 @@ package moze_intel.projecte.gameObjs.blocks;
 
 import cpw.mods.fml.common.network.NetworkRegistry;
 import moze_intel.projecte.PECore;
-import moze_intel.projecte.api.IPedestalItem;
+import moze_intel.projecte.api.item.IPedestalItem;
 import moze_intel.projecte.gameObjs.ObjHandler;
 import moze_intel.projecte.gameObjs.tiles.DMPedestalTile;
 import moze_intel.projecte.gameObjs.tiles.TileEmc;
