+++ src/main/java/moze_intel/projecte/gameObjs/items/AlchemicalBag.java
@@ -3,7 +3,7 @@ package moze_intel.projecte.gameObjs.items;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import moze_intel.projecte.PECore;
-import moze_intel.projecte.api.IAlchBagItem;
+import moze_intel.projecte.api.item.IAlchBagItem;
 import moze_intel.projecte.gameObjs.ObjHandler;
 import moze_intel.projecte.gameObjs.container.AlchBagContainer;
 import moze_intel.projecte.playerData.AlchemicalBags;
