+++ src/main/java/moze_intel/projecte/gameObjs/items/CataliticLens.java
@@ -2,7 +2,7 @@ package moze_intel.projecte.gameObjs.items;
 
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
-import moze_intel.projecte.api.IProjectileShooter;
+import moze_intel.projecte.api.item.IProjectileShooter;
 import moze_intel.projecte.gameObjs.entity.EntityLensProjectile;
 import moze_intel.projecte.utils.Constants;
 import net.minecraft.client.renderer.texture.IIconRegister;
