+++ src/main/java/moze_intel/projecte/gameObjs/items/MercurialEye.java
@@ -3,7 +3,7 @@ package moze_intel.projecte.gameObjs.items;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import moze_intel.projecte.PECore;
-import moze_intel.projecte.api.IExtraFunction;
+import moze_intel.projecte.api.item.IExtraFunction;
 import moze_intel.projecte.utils.Constants;
 import moze_intel.projecte.utils.EMCHelper;
 import net.minecraft.block.Block;
