+++ src/main/java/moze_intel/projecte/gameObjs/items/PhilosophersStone.java
@@ -4,8 +4,8 @@ import cpw.mods.fml.common.network.NetworkRegistry.TargetPoint;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import moze_intel.projecte.PECore;
-import moze_intel.projecte.api.IExtraFunction;
-import moze_intel.projecte.api.IProjectileShooter;
+import moze_intel.projecte.api.item.IExtraFunction;
+import moze_intel.projecte.api.item.IProjectileShooter;
 import moze_intel.projecte.gameObjs.ObjHandler;
 import moze_intel.projecte.gameObjs.entity.EntityMobRandomizer;
 import moze_intel.projecte.gameObjs.tiles.TileEmc;
