+++ src/main/java/moze_intel/projecte/gameObjs/items/VolcaniteAmulet.java
@@ -8,9 +8,9 @@ import com.google.common.collect.Lists;
 import cpw.mods.fml.common.Optional;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
-import moze_intel.projecte.api.IFireProtectionItem;
-import moze_intel.projecte.api.IPedestalItem;
-import moze_intel.projecte.api.IProjectileShooter;
+import moze_intel.projecte.utils.IFireProtectionItem;
+import moze_intel.projecte.api.item.IPedestalItem;
+import moze_intel.projecte.api.item.IProjectileShooter;
 import moze_intel.projecte.config.ProjectEConfig;
 import moze_intel.projecte.gameObjs.entity.EntityLavaProjectile;
 import moze_intel.projecte.handlers.PlayerChecks;
