+++ src/main/java/moze_intel/projecte/gameObjs/items/rings/Arcana.java
@@ -5,7 +5,6 @@ import java.util.List;
 import baubles.api.BaubleType;
 import baubles.api.IBauble;
 import cpw.mods.fml.common.Optional;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.texture.IIconRegister;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
@@ -20,11 +19,11 @@ import net.minecraft.util.IIcon;
 import net.minecraft.util.MathHelper;
 import net.minecraft.util.StatCollector;
 import net.minecraft.world.World;
-import moze_intel.projecte.api.IExtraFunction;
-import moze_intel.projecte.api.IFireProtectionItem;
-import moze_intel.projecte.api.IFlightItem;
-import moze_intel.projecte.api.IModeChanger;
-import moze_intel.projecte.api.IProjectileShooter;
+import moze_intel.projecte.api.item.IExtraFunction;
+import moze_intel.projecte.utils.IFireProtectionItem;
+import moze_intel.projecte.utils.IFlightItem;
+import moze_intel.projecte.api.item.IModeChanger;
+import moze_intel.projecte.api.item.IProjectileShooter;
 import moze_intel.projecte.gameObjs.entity.EntityFireProjectile;
 import moze_intel.projecte.gameObjs.entity.EntityLightningProjectile;
 import moze_intel.projecte.gameObjs.items.ItemPE;
