+++ src/main/java/moze_intel/projecte/gameObjs/items/rings/Ignition.java
@@ -6,12 +6,11 @@ import baubles.api.IBauble;
 import com.google.common.collect.Lists;
 
 import cpw.mods.fml.common.Optional;
-import moze_intel.projecte.api.IFireProtectionItem;
-import moze_intel.projecte.api.IPedestalItem;
-import moze_intel.projecte.api.IProjectileShooter;
+import moze_intel.projecte.utils.IFireProtectionItem;
+import moze_intel.projecte.api.item.IPedestalItem;
+import moze_intel.projecte.api.item.IProjectileShooter;
 import moze_intel.projecte.config.ProjectEConfig;
 import moze_intel.projecte.gameObjs.entity.EntityFireProjectile;
-import moze_intel.projecte.gameObjs.entity.EntityLightningProjectile;
 import moze_intel.projecte.gameObjs.tiles.DMPedestalTile;
 import moze_intel.projecte.handlers.PlayerChecks;
 import moze_intel.projecte.utils.MathUtils;
@@ -23,8 +22,6 @@ import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
-import net.minecraft.entity.projectile.EntitySnowball;
-import net.minecraft.init.Blocks;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.DamageSource;
 import net.minecraft.util.EnumChatFormatting;
