+++ src/main/java/moze_intel/projecte/utils/IFlightItem.java
@@ -1,4 +1,4 @@
-package moze_intel.projecte.api;
+package moze_intel.projecte.utils;
 
 public interface IFlightItem // Marker interface for PlayerChecks
 {
